DEBUG: quic: add a very simple Rx perturbation mechanism
When HA_QUIC_RAND_RECV is set at build time, some code is added to memorize
incoming packets int an array of buffers. Then based on a random value that
is compared to environment variable HA_QUIC_RAND_RECV_RATE (0..100), instead
of delivering the received packet, one of the 10 other packets (or none) is
returned. This is very limited but is sufficient as a first step to mimmick
duplication, reordering and loss.

Note that since the environment variable parser is called late, it is also
possible to add this to the global section:

      presetenv HA_QUIC_RAND_RECV_RATE 10

for example to require that 10% of incoming packets are perturbed.